% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acstp.R
\name{ACSTP}
\alias{ACSTP}
\alias{acstp}
\title{Function for fNIRS signal extraction}
\usage{
acstp(dat, dots, n.c, time, threshold = NULL, quali = NULL)
}
\arguments{
\item{dat}{array (time X channel X trial) containig all the trials}

\item{dots}{array (time X channel X trial) containing the 
trials of z class}

\item{n.c}{names of the rows with which build mask up, 
for the automatic choice of the subspace dimension}

\item{time}{names of the columns with which build mask up, for 
the automatic choice of the subspace dimension}

\item{threshold}{default: the data are cleaned choosing the eigenvalues and
         eigenvectors of spatial and temporal variance and
         covariance matrix that are bigger than a threshold
         equal to (1/10^6)* 1ü¾Œ†ˆ¼ eigenvalues of the spatial 
         variance and covariance matrix
if threshold = "autovalori": only the eigenvalues and the corresponding
                             eigenvectors from the first
                             to that specified by the parameter "quali"
                             are kept}

\item{quali}{specifies the last eigenvalue and corresponding eigenvector
of the spatial and temporal covariance matrix 
that are to be kept starting from the first.
quali is not NULL only when threshold is equal to "autovalori"}
}
\value{
a list containing in this order:
        p subspace dimension
        Q matrix containing the signal extracted with the function
        weights a vector of weights for evry trials 
        SNR a number that indicate the SNR of the extracted signal in dB
}
\description{
(English)
             Function that extract signal from fNIRS data, using 
             a spatio-temporal filter made up of a pair of matrices
             whose number of columns is that that maximize the SNR of
             the class under analysis.
             A class corresponds to a specific stimulus.
             The pair of matrices is achieved by a two step procedure
             1° step--> denoising of the data with a pair of matrices
                        make up of eigenvalues and egeinvectors of the
                        spatial and temporal variance and covariance matrix
                        of the data. Applying this pair of matrices 
                        to the mean of the trial of the z class, you get
                        clean data
             2° step--> In this step we apply the singular values 
                        decomposition to the clean data obtained in the 
                        previous step. The number of the components, that
                        are found through the svd that are to be keep, is 
                        automatically 
                        chosen from the function in a way to maximize
                        the SNR of the class under analysis
             (Italian)           
             Funzione che estrae il segnale da dati fNIRS 
             eseguendo un filtraggio spazio-temporale servendosi di 
             una coppia di matrici il cui numero di colonne 
             è scelto automaticamente dalla funzione in
             modo da massimizzare il SNR della classe in esame.
             La classe corrisponde a un determinato stimolo.
             La coppia di matrici è costruita secondo una procedura
             a due passi:
             1° passo --> fase di denoising dei dati attraverso 
                          la costruzione di due matrici ottenute 
                          a seguito della decomposizione spettrali
                          delle matrici di varianza e covarianza
                          spaziale e temporale dei dati. Ottenute tali 
                          matrici, si applicano alla media dei 
                          trials della classe z.
              2° passo--> estrazione dell'informazione dai dati puliti 
                          al passo 1°. La media filtrata al passo 1°
                          viene decomposta con il metodo di 
                          decomposizione ai valori singolari. Il
                          numero di componenti ottenute dalla svd 
                          da tenere, è scelto in maniera automatica 
                          dall'algoritmo. Tale numero è scelto 
                          in maniera tale da massimizzare il
                          SNR della classe in esame.
}
\author{
Michelle Viola
}
\seealso{
link https://sites.google.com/site/marcocongedo
}

