% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neuR_funct.R
\name{add.neuR.funct}
\alias{add.neuR.funct}
\title{add funct to neuR-object}
\usage{
add.neuR.funct(obj, funct, out.name = funct, ...)
}
\arguments{
\item{obj}{neuR-object}

\item{out.name}{map by default. name of the function to be added in obj@function$out.name}

\item{map}{function to be returned}
}
\value{
a neuR-object
}
\description{
add funct to neuR-object
}
