% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neuR_map.R
\name{add.neuR.map}
\alias{add.neuR.map}
\title{add volume to neuR-object}
\usage{
add.neuR.map(obj, map, out.name = map, ...)
}
\arguments{
\item{obj}{neuR-object}

\item{map}{function to be returned}

\item{out.name}{map by default. name of the map to be added in obj@function$out.name}
}
\value{
a neuR-object
}
\description{
add volume to neuR-object
}

