% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_irc.R
\name{compute.irc}
\alias{compute.irc}
\title{Computes IRC (intra run correlation) of a neuR-object}
\usage{
compute.irc(D, pc.num = 1, drop.tcs = FALSE)
}
\arguments{
\item{D}{a neuR-ogject}

\item{pc.num}{1}

\item{drop.tcs}{FALSE}
}
\value{
a 3D array
}
\description{
Computes IRC (intra run correlation) of a neuR-object pcs array of slot data of a neuR-object (i.e. D@data$pcs)
}

