% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_irh.R
\name{compute.irh}
\alias{compute.irh}
\title{Computes IRH (intra run homogeneity) of a neuR-object}
\usage{
compute.irh(D, pc.num = 1, drop.tcs = FALSE)
}
\arguments{
\item{D}{a neuR-object}

\item{pc.num}{1}

\item{drop.tcs}{FALSE}
}
\value{
a 3D array
}
\description{
Computes IRH (intra run homogeneity) of a neuR-object pcs array of slot data of a neuR-object (i.e. D@data$pcs)
indice di omogeneità dei k loadings. indice in scala 0-1. 
la norma essendo pari a 1, implica un momento secondo pari a 1/k.
il momento secondo è decomponibile in varianza + media^2.
l'indice è pari a media(loadings)^2*k.
se 1 tutti i loading sono uguali tra loro (e pari a 1/sqrt(k))
se 0, i loadings hanno media 0 e la varianza vale 1/k
}
