% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pcs.R
\name{compute.pcs}
\alias{compute.pcs}
\title{Computes principal components for each voxel/channel of a neuR-object}
\usage{
compute.pcs(D, center = TRUE, scale = FALSE, max.pc.num = 1,
  drop.tcs = TRUE, selected.volumes = NULL, ...)
}
\arguments{
\item{D}{a neuR-ogject}

\item{center}{TRUE}

\item{scale}{FALSE}

\item{max.pc.num}{1}

\item{drop.tcs}{TRUE}

\item{selected.volumes}{NULL or a logical vector of length n (number of volumes per block)}

\item{...}{other}
}
\value{
a neuR-object
}
\description{
Computes principal components for each voxel/channel stored in tcs array of slot data of a neuR-object (i.e. D@data$tcs)
}

