% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Test.R
\name{compute.test}
\alias{compute.test}
\title{Computes Tests in a neuR-object}
\usage{
compute.test(D, pc.num = 1, drop.tcs = FALSE)
}
\arguments{
\item{D}{a neuR-ogject}

\item{pc.num}{1}

\item{drop.tcs}{FALSE}
}
\value{
a 3D array
}
\description{
Computes tests of a neuR-object pcs array of slot data of a neuR-object (i.e. D@data$pcs)
}

