% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weightedPT.R
\name{compute_weighted_P}
\alias{compute_weighted_P}
\title{Computes weighted p-values (and T) of a neuR-object}
\usage{
compute_weighted_P(D, Pmap = "test.P", Wmap = "IRC", returnMaps = "P",
  names_out = c(Pmap = paste("test.P_W", Wmap, sep = ""), Zmap =
  paste("test.T_W", Wmap, sep = "")), tail = NULL, SignMap = "test.T")
}
\arguments{
\item{D}{a neuR-object}

\item{Pmap}{charachter (i.e. the name of the map in D) or an array. "test.P" by default.}

\item{Wmap}{charachter (i.e. the name of the map in D) or an array. "IRC" by default.}

\item{tail}{the sign of the alternative. It can be 0, 1, -1.}

\item{SignMap}{(only needed if tail=0) the name of the map that has the sign of the test statistic. "test.T" by default}

\item{return}{can be "Pmap", "Zmap" or c("Pmap","Zmap")}
}
\value{
Maps a named list of array/s
}
\description{
Computes IRC (intra run correlation) of a neuR-object pcs array of slot data of a neuR-object (i.e. D@data$pcs)
}

