% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_neuR.R
\name{dim.map}
\alias{dim.map}
\title{dim of a given map}
\usage{
\method{dim}{map}(object, map.name)
}
\arguments{
\item{object}{a neuR-object}

\item{map.name}{the map to know the dims}
}
\value{
a vector of dims: n replicates X nvoxels X n dimensions
}
\description{
dim of a given map
}

