% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vxls.R
\name{extractVxls}
\alias{extractVxls}
\alias{extract_tc}
\title{extractVxls}
\usage{
extractVxls(xyz, D)
}
\arguments{
\item{xyz}{a nX3 matrix of n coordinates}

\item{D}{a neuR-object}

\item{tcs}{any proper array e.g. from D@data$tcs}

\item{mask}{usually D@mask}
}
\description{
select voxels (or time courses of voxels) 
from a matrix of coordinates
}

