% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neuR_funct.R, R/get_neuR_map.R
\name{get.neuR.map}
\alias{get.neuR.map}
\alias{get.neuR.funct}
\alias{get.neuR.map}
\title{get functions to compute maps neuR-object}
\usage{
get.neuR.funct(funct, ...)

get.neuR.map(obj, map, recompute = FALSE, ...)
}
\arguments{
\item{funct}{}

\item{...}{are used to set the parameters of the map}

\item{obj}{neuR-object}

\item{map}{function name to be returned or the array itself}

\item{recompute}{FALSE}
}
\value{
a function

a volume
}
\description{
NOT IMPLEMENTED YET: out.funct.name=map the name of the function to be generated (eg different names depending on the parameters)

NOT IMPLEMENTED YET: out.funct.name=map the name of the function to be generated (eg different names depending on the parameters)
}
