% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_neuR.R
\name{names.map}
\alias{names.map}
\title{dimnames of the maps}
\usage{
\method{names}{map}(object)
}
\arguments{
\item{object}{a neuR-object}
}
\value{
a list of dimnames
}
\description{
dimnames of the maps
}
