% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuR-object.R
\docType{class}
\name{neuR.object-class}
\alias{neuR.object-class}
\title{An S4 class to store fNIRS, fMRI, EEG data}
\description{
An S4 class to store fNIRS, fMRI, EEG data
}
\section{Slots}{

\describe{
\item{\code{data}}{list of 3D arrays}

\item{\code{mask}}{}

\item{\code{info}}{}
}}

