% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuR
\docType{package}
\name{neuR}
\alias{neuR}
\alias{neuR-package}
\title{neuR: R package for neuroscience data processing and statistical analysis of neuro-data}
\description{
The neuR package provides functions to deal with fMRI, fNIRS, EEG data.
}
\section{neuR functions}{

}

