% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_neuR.R
\name{nvoxels}
\alias{nvoxels}
\title{Number of voxels of an neuR-object}
\usage{
nvoxels(object)
}
\arguments{
\item{object}{a neuR-object}
}
\value{
Number of voxels of an neuR-object, i.e. a scalar
}
\description{
Number of voxels of an neuR-object
}

