% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vxls.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(D, reduce.by = 2, summary_function = rowMeans)
}
\arguments{
\item{D}{a neuR-object}

\item{reduce.by}{(integer or vector) resizing factor.}

\item{summary_function}{by default: \code{rowMeans} 
If it is a vector, the tree coordinates indicate the 
rescaling to bi applied to each dimension. 
If it is a scalar, the same rescaling is applied to the 3 dims.}
}
\description{
reduce a neuR-object
}
