% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_map.R
\name{summary.map}
\alias{summary.map}
\title{summary of all maps}
\usage{
summary.map(object, maps = NULL)
}
\arguments{
\item{object}{a neuR-object}

\item{maps}{NULL (default) or a vecotr of names of maps}
}
\value{
a table of summary stats for each volume
}
\description{
summary of all maps
}
