% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.volumes.R
\name{write.volumes}
\alias{write.volumes}
\title{writes  .nii and .img/.hdr file file from a neuR-object}
\usage{
write.volumes(D, which.maps = "tcs", file.root.name = "V",
  file.type = "nifti", into.path = ".")
}
\arguments{
\item{D}{neuR-object}

\item{which.maps}{"tcs"}

\item{file.root.name}{"V"}

\item{file.type}{"nifti"}

\item{into.path}{"."}
}
\description{
writes  .nii and .img/.hdr file file from a neuR-object
}

