% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neuR_map.R
\name{add.neuR.map}
\alias{add.neuR.map}
\title{add volume to neuR-object}
\usage{
add.neuR.map(obj, map, out.name = NULL, ...)
}
\arguments{
\item{obj}{neuR-object}

\item{map}{name of the map (function) to be added. It can also be a list of maps (i.e. 3D arrays)}

\item{out.name}{equal to map (or names(map) if map is a list) by default. 
Name of the map to be added to the returned object.}
}
\value{
a neuR-object
}
\description{
add volume to neuR-object
}
