% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test.R
\name{compute.test}
\alias{compute.test}
\title{Computes Tests in a neuR-object}
\usage{
compute.test(D, left.array, right.formula = ~1, offset.values = NULL,
  tail = 1)
}
\arguments{
\item{D}{a neuR-object}

\item{left.array}{the name of the map in D@data to be used in the model or the array itself}

\item{right.formula}{~1 by default. NOT IMPLEMENTED YET. The names refer to the columnames of D@info$design}

\item{offset.values}{(NULL by default, same effect as 0) A value to be subtracted to each value of the matrix (same dims of left.array)
It may also be 'meanOverall' or 'meanBySubject' (i.e. average over the second dimension)}
}
\value{
a list of 3D arrays, usually a T and P one.
}
\description{
Computes tests of a neuR-object pcs array of slot data of a neuR-object (i.e. D@data$pcs)
}

