% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vxls.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(D, reduce.by = 2)
}
\arguments{
\item{D}{a neuR-object}

\item{reduce.by}{(integer) resizing factor. 
If it is a vector, the tree coordinates indicate the 
rescaling to bi applied to each dimension. 
If it is a scalar, the same rescaling is applied to the 3 dims.}
}
\description{
reduce a neuR-object
}

