% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fMRI.data.R
\name{read.fMRI.data}
\alias{read.fMRI.data}
\title{Reads .nii and .img/.hdr file data and save to neuR-object}
\usage{
read.fMRI.data(path = ".", pattern = "s.*\\\\.img", files = NULL,
  mask = NULL, info = NULL, silent = FALSE, exclude.files = c())
}
\arguments{
\item{path}{"."}

\item{pattern}{"s.*\\.img"}

\item{files}{NULL}

\item{mask}{NULL}

\item{info}{NULL}

\item{silent}{FALSE}

\item{exclude.files}{vector of ids of files to exclude}
}
\value{
a neuR-object
}
\description{
Reads .nii and .img/.hdr file data and save to neuR-object
}

