% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fMRI.data.R
\name{read.fMRI.data}
\alias{read.fMRI.data}
\title{Reads .nii and .img/.hdr file data and save to neuR-object}
\usage{
read.fMRI.data(path = ".", pattern = "s.*\\\\.img", files = NULL,
  mask = "constant", info = NULL, silent = FALSE, exclude.files = c(),
  header.file = NULL)
}
\arguments{
\item{path}{"."}

\item{pattern}{"s.*\\.img"}

\item{files}{NULL}

\item{mask}{a file name, a array, equal to 'constant' (look for non constant voxels) 
or a scalar (values equal to this number are out of the brain)}

\item{info}{NULL}

\item{silent}{FALSE}

\item{exclude.files}{vector of ids of files to exclude}

\item{header.file}{name of the file from which the head should be read. if NULL (default) the head of the first file (not excluded by exclude.files) is used}
}
\value{
a neuR-object
}
\description{
Reads .nii and .img/.hdr file data and save to neuR-object
}

